program-error: Remove solana-program usage
#### Problem

A few SPL programs use `spl-program-error`, but it brings in
solana-program by default, which makes program-error definitions very
heavy to include.

#### Summary of changes

Split up the usage of `solana-program-error` and `solana-decode-error`,
parametrize them just like `solana-program` used to be.

As part of this change, I *didn't* parametrize `solana-msg`, since that
just resolves to a simple syscall. We could inline the syscall and avoid
its usage entirely, but that seems a bit harsh.

With this change, we go from building 192 crates to 40, and build times
go from ~6.5s to ~2.2s on my machine.

NOTE: This is a breaking change.