CI: Pin cargo-release and cargo-semver-checks (#129)
#### Problem

The release job is failing because it can't install cargo-semver-checks
due to the lower Rust version declared in the repo.

Also, cargo-release just bumped up to 1.87 on master, which means that a
release with a higher MSRV is just around the corner.

#### Summary of changes

Pin the versions of both tools to the latest that supports Rust 1.86.