CI: Only run semver-checks for lib crates
#### Problem

A publish failed on a proc-macro crate because semver-checks doesn't
work on non-lib crates.

#### Summary of changes

Be a bit smarter, and only run semver-checks if the package has a
crate-type that contains the word "lib", which means it'll catch
`cdylib`, `rlib`, and `lib`.

At the same time, I removed `cdylib` from a few crates that didn't need
it, and added a lib target to those that didn't explicitly declare it.