CI: Don't specify bash for the semver step (#135)
#### Problem

The publish job fails for proc-macro crates because they don't specify a
crate-type. This means that the first call to `toml get ...` fails,
which causes the whole job to fail.

#### Summary of changes

Cargo issues a warning if you specify `crate-type = ["proc-macro"]`, so
rather than creating a new warning, simply allow for errors when
searching for the string.