Move repo to use a Makefile
#### Problem

The libraries repo has a lot of packages to run CI and commands against,
and contains a pretty long list of commands in package.json, so it's a
great candidate for moving to a Makefile.

#### Summary of changes

This is pretty similar to the others, except that I removed the "first
word before - means directory" and instead flatten everything at the
top-level.